# airPromo/src/infra/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentação Swagger do airPromo",
        "version": "1.0.0",
        "title": "airPromo",
        "contact": {
            "email": "keltonmof@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "schemes": [
        "http"
    ],
    "host": "189.48.177.66:3000",
    "basePath": "/",
    "paths": {
        "/resposta": {
            "post": {
                "summary": "Envia uma resposta",
                "tags": ["Respostas"],
                "description": "Envia uma resposta a ser salva para mostrar no painel e nos relatórios",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Objeto Resposta",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/createRespostaPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resposta criada",
                        "schema": {
                            "$ref": "#/definitions/createRespostaResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    },
                    "422": {
                        "description": "Parâmetro inválido",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    },
                    "500": {
                        "description": "Erro no servidor de banco de dados",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/respostas/quantidade": {
            "get": {
                "summary": "Quantidade de registros por resposta",
                "tags": ["Respostas"],
                "description": "Lista quantas vezes cada resposta foi registrada para se calcular o tamanho na tela",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Dados recuperados com sucesso",
                        "schema": {
                            "$ref": "#/definitions/getRespostasQuantidadesResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    },
                    "500": {
                        "description": "Erro no servidor de banco de dados",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/respostas/novos": {
            "get": {
                "summary": "Respostas ainda não retornadas.",
                "tags": ["Respostas"],
                "description": "Lista as respostas que ainda não foram enviadas na requisição.",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Dados recuperados com sucesso",
                        "schema": {
                            "$ref": "#/definitions/getRespostasNovasResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    },
                    "500": {
                        "description": "Erro no servidor de banco de dados",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/respostas/relatorio-tematico": {
            "get": {
                "summary": "Relatório excel temático",
                "tags": ["Respostas"],
                "description": "Relatório em excel com os temas utilizados nas respostas, as respostas e quantidades",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Relatório gerado com sucesso",
                        "schema": {
                            "$ref": "#/definitions/getRelatorioResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    },
                    "500": {
                        "description": "Erro no servidor de banco de dados",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/respostas/relatorio-percentual": {
            "get": {
                "summary": "Relatório excel percentual",
                "tags": ["Respostas"],
                "description": "Relatório em excel com os temas utilizados nas respostas e seus percentuais",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Relatório gerado com sucesso",
                        "schema": {
                            "$ref": "#/definitions/getRelatorioResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    },
                    "500": {
                        "description": "Erro no servidor de banco de dados",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/temas/quantidade": {
            "get": {
                "summary": "Percentual dos temas utilizados",
                "tags": ["Temas"],
                "description": "Os temas respondidos e seus percentuais",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Termas obtidos com sucesso",
                        "schema": {
                            "$ref": "#/definitions/getTemasQuantidadeResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    },
                    "500": {
                        "description": "Erro no servidor de banco de dados",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "InvalidResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "login": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "loginResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "createRespostaPayload": {
            "type": "object",
            "properties": {
                "Texto": {
                    "type": "string"
                }
            }
        },
        "createRespostaResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer"
                },
                "CodigoResposta": {
                    "type": "integer"
                }
            }
        },
        "getRespostasQuantidadesResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "Texto": {
                                "type": "string"
                            },
                            "Tema": {
                                "type": "string"
                            },
                            "Quantidade": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "getRelatorioResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer"
                },
                "data": {
                    "type": "object",
                    "properties": {
                        "filename": {
                            "type": "string"
                        },
                        "content": {
                            "type": "string"
                        },
                        "encoding": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "getRespostasNovasResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "Texto": {
                                "type": "string"
                            },
                            "Retonarda": {
                                "type": "string"
                            },
                            "CodigoResposta": {
                                "type": "integer"
                            }
                        }
                    }
                }
            }
        },
        "getTemasQuantidadeResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "Tema": {
                                "type": "string"
                            },
                            "PercentualCategoria": {
                                "type": "number"
                            },
                            "PercentualResposta": {
                                "type": "number"
                            }
                        }
                    }
                }
            }
        }
    }
}